archive/pydantic_inference/requirements.txt
pydantic-ai
google-generativeai
google-cloud-aiplatform
python-dotenv
PyMuPDF
pydantic
Pillow